Better ‘simple’ example, showing how to start consumer in thread.
In real world you might have more than 1 channels open, each running
consumers. Because start_consuming method blocks, we need to run it
in thread.
Currently there is no way to stop the consumer, so in this example
we’re simply joining the consumer thread, but it will never finish.