Properly handle full MPSC channels
When using SyncSender, calling send() will
block until space is available for the new
message. When combined with other issues (for
example: mandatory message deliveries failing
and the client not immediately handling the
error replies) a channel's SyncSender channel
can easily fill up causing deadlocks in a
consumer / producer.

Always first try to do a non-blocking send. If
this fails, warn the channel is full and try
again with a blocking send. This retains the
existing behavior, but includes better output
for debugging why your program is deadlocking.